Use explicit resource layouts for fixed function pipelines
