+++ src/OpenSage.Game/Settings/LightSettings.cs
@@ -4,12 +4,12 @@ namespace OpenSage.Settings
 {
     public sealed class LightSettings
     {
-        internal readonly GlobalLightingTypes.LightingConstantsPS TerrainLightsPS;
-        internal readonly GlobalLightingTypes.LightingConstantsPS ObjectLightsPS;
+        internal readonly GlobalShaderResources.LightingConstantsPS TerrainLightsPS;
+        internal readonly GlobalShaderResources.LightingConstantsPS ObjectLightsPS;
 
         internal LightSettings(
-            in GlobalLightingTypes.LightingConstantsPS terrainLightsPS,
-            in GlobalLightingTypes.LightingConstantsPS objectLightsPS)
+            in GlobalShaderResources.LightingConstantsPS terrainLightsPS,
+            in GlobalShaderResources.LightingConstantsPS objectLightsPS)
         {
             TerrainLightsPS = terrainLightsPS;
             ObjectLightsPS = objectLightsPS;
